Rename and fix issues with missing Api Key
