+++ AzzyBot-Next/Services/Modules/AzuraCastService.cs
@@ -4,7 +4,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using AzzyBot.Utilities.Records.AzuraCast;
 
-namespace AzzyBot.Services;
+namespace AzzyBot.Services.Modules;
 
 public sealed class AzuraCastService(WebRequestService webService)
 {
@@ -21,6 +21,14 @@ public sealed class AzuraCastService(WebRequestService webService)
             throw new InvalidOperationException($"API response is empty, url: {url}");
 
         return body;
+    }
+
+    private static Dictionary<string, string> CreateHeader(string apiKey)
+    {
+        return new()
+        {
+            ["X-API-Key"] = apiKey
+        };
     }
 
     private async Task<T> FetchFromApiAsync<T>(Uri baseUrl, string endpoint, Dictionary<string, string>? headers = null)
@@ -44,22 +52,24 @@ public sealed class AzuraCastService(WebRequestService webService)
         return JsonSerializer.Deserialize<List<T>>(body) ?? throw new InvalidOperationException($"Could not deserialize body: {body}");
     }
 
-    public Task<PlaylistRecord> GetPlaylistAsync(Uri baseUrl, int stationId, int playlistId)
+    public Task<PlaylistRecord> GetPlaylistAsync(Uri baseUrl, string apiKey, int stationId, int playlistId)
     {
+        ArgumentException.ThrowIfNullOrWhiteSpace(apiKey, nameof(apiKey));
         ArgumentOutOfRangeException.ThrowIfNegativeOrZero(stationId, nameof(stationId));
 
         string endpoint = $"station/{stationId}/playlist/{playlistId}";
 
-        return FetchFromApiAsync<PlaylistRecord>(baseUrl, endpoint);
+        return FetchFromApiAsync<PlaylistRecord>(baseUrl, endpoint, CreateHeader(apiKey));
     }
 
-    public Task<IReadOnlyList<PlaylistRecord>> GetPlaylistsAsync(Uri baseUrl, int stationId)
+    public Task<IReadOnlyList<PlaylistRecord>> GetPlaylistsAsync(Uri baseUrl, string apiKey, int stationId)
     {
+        ArgumentException.ThrowIfNullOrWhiteSpace(apiKey, nameof(apiKey));
         ArgumentOutOfRangeException.ThrowIfNegativeOrZero(stationId, nameof(stationId));
 
         string endpoint = $"station/{stationId}/playlists";
 
-        return FetchFromApiListAsync<PlaylistRecord>(baseUrl, endpoint);
+        return FetchFromApiListAsync<PlaylistRecord>(baseUrl, endpoint, CreateHeader(apiKey));
     }
 
     public Task<NowPlayingDataRecord> GetNowPlayingAsync(Uri baseUrl, int stationId)
