+++ AzzyBot-Next/Utilities/Records/AzuraCast/PlaylistRecord.cs
